Refactor enums to use byte as underlying type; enhance documentation for TaskPriority and TaskStatus; introduce EOrganizationUserLevel and ETaskUserLevel enums

+++ src/QFXTaskMan.Core/Enumerables/ELogAction.cs
@@ -1,6 +1,6 @@
 namespace QFXTaskMan.Core.Enumerables;
 
-public enum ELogAction : int
+public enum ELogAction : byte
 {
     Create = 1,
     Update = 2,
